code reuse instead of duplication
