+++ src/constants.ts
@@ -1,0 +1,2 @@
+export const csnInteropEffectiveV1ExtensionsDocumentationPage =
+  "https://sap.github.io/csn-interop-specification/spec-v1/extensions/";
